+++ native/c/zusf.cpp
@@ -838,7 +838,7 @@ string zusf_build_mode_string(mode_t mode)
  *
  * @return RTNCD_SUCCESS on success, RTNCD_FAILURE on failure
  */
-int zusf_create_uss_file_or_dir(ZUSF *zusf, string file, mode_t mode, bool createDir)
+int zusf_create_uss_file_or_dir(ZUSF *zusf, const string &file, mode_t mode, bool createDir)
 {
   struct stat file_stats;
   if (stat(file.c_str(), &file_stats) != -1)
@@ -1044,7 +1044,7 @@ int zusf_list_uss_file_path(ZUSF *zusf, string file, string &response, ListOptio
  *
  * @return RTNCD_SUCCESS on success, RTNCD_FAILURE on failure
  */
-int zusf_read_from_uss_file(ZUSF *zusf, string file, string &response)
+int zusf_read_from_uss_file(ZUSF *zusf, const string &file, string &response)
 {
   const auto bpxk_autocvt = getenv("_BPXK_AUTOCVT");
   setenv("_BPXK_AUTOCVT", "OFF", 1);
@@ -1073,17 +1073,20 @@ int zusf_read_from_uss_file(ZUSF *zusf, string file, string &response)
 
   if (zusf->encoding_opts.data_type == eDataTypeText)
   {
-    // Try to get the file's CCSID first
-    int file_ccsid = zusf_get_file_ccsid(zusf, file);
-    if (file_ccsid > 0 && file_ccsid != 65535) // Valid CCSID and not binary
-    {
-      encoding_to_use = zut_int_to_string(file_ccsid);
-      has_encoding = true;
-    }
-    else if (strlen(zusf->encoding_opts.codepage) > 0)
+    if (strlen(zusf->encoding_opts.codepage) > 0)
     {
       encoding_to_use = string(zusf->encoding_opts.codepage);
       has_encoding = true;
+    }
+    else
+    {
+      // Use tagged encoding if valid CCSID and not UTF-8 or binary
+      int file_ccsid = zusf_get_file_ccsid(zusf, file);
+      if (file_ccsid > 0 && file_ccsid != 1208 && file_ccsid != 65535)
+      {
+        encoding_to_use = zut_int_to_string(file_ccsid);
+        has_encoding = true;
+      }
     }
   }
 
@@ -1114,11 +1117,18 @@ int zusf_read_from_uss_file(ZUSF *zusf, string file, string &response)
  * @param zusf pointer to a ZUSF object
  * @param file name of the USS file
  * @param pipe name of the output pipe
+ * @param content_len pointer where the length of the file contents will be stored
  *
  * @return RTNCD_SUCCESS on success, RTNCD_FAILURE on failure
  */
-int zusf_read_from_uss_file_streamed(ZUSF *zusf, string file, string pipe)
+int zusf_read_from_uss_file_streamed(ZUSF *zusf, const string &file, const string &pipe, size_t *content_len)
 {
+  if (content_len == nullptr)
+  {
+    zusf->diag.e_msg_len = sprintf(zusf->diag.e_msg, "content_len must be a valid size_t pointer");
+    return RTNCD_FAILURE;
+  }
+
   FILE *fin = fopen(file.c_str(), zusf->encoding_opts.data_type == eDataTypeBinary ? "rb" : "r");
   if (!fin)
   {
@@ -1140,17 +1150,20 @@ int zusf_read_from_uss_file_streamed(ZUSF *zusf, string file, string pipe)
 
   if (zusf->encoding_opts.data_type == eDataTypeText)
   {
-    // Try to get the file's CCSID first
-    int file_ccsid = zusf_get_file_ccsid(zusf, file);
-    if (file_ccsid > 0 && file_ccsid != 65535) // Valid CCSID, not binary
-    {
-      encoding_to_use = zut_int_to_string(file_ccsid);
-      has_encoding = true;
-    }
-    else if (strlen(zusf->encoding_opts.codepage) > 0)
+    if (strlen(zusf->encoding_opts.codepage) > 0)
     {
       encoding_to_use = string(zusf->encoding_opts.codepage);
       has_encoding = true;
+    }
+    else
+    {
+      // Use tagged encoding if valid CCSID and not UTF-8 or binary
+      int file_ccsid = zusf_get_file_ccsid(zusf, file);
+      if (file_ccsid > 0 && file_ccsid != 1208 && file_ccsid != 65535)
+      {
+        encoding_to_use = zut_int_to_string(file_ccsid);
+        has_encoding = true;
+      }
     }
   }
 
@@ -1179,8 +1192,15 @@ int zusf_read_from_uss_file_streamed(ZUSF *zusf, string file, string pipe)
       }
     }
 
-    temp_encoded = zbase64::encode(chunk, chunk_len);
+    *content_len += chunk_len;
+    temp_encoded = zbase64::encode(chunk, chunk_len, false);
     fwrite(&temp_encoded[0], 1, temp_encoded.size(), fout);
+  }
+
+  const auto padding = 4 - (*content_len % 4);
+  if (padding > 0)
+  {
+    fwrite("===", 1, padding, fout);
   }
 
   fflush(fout);
@@ -1199,7 +1219,7 @@ int zusf_read_from_uss_file_streamed(ZUSF *zusf, string file, string pipe)
  *
  * @return RTNCD_SUCCESS on success, RTNCD_FAILURE on failure
  */
-int zusf_write_to_uss_file(ZUSF *zusf, string file, string &data)
+int zusf_write_to_uss_file(ZUSF *zusf, const string &file, string &data)
 {
   // TODO(zFernand0): Avoid overriding existing files
   struct stat file_stats;
@@ -1224,17 +1244,20 @@ int zusf_write_to_uss_file(ZUSF *zusf, string file, string &data)
 
   if (zusf->encoding_opts.data_type == eDataTypeText)
   {
-    // Try to get the file's CCSID first
-    int file_ccsid = zusf_get_file_ccsid(zusf, file);
-    if (file_ccsid > 0 && file_ccsid != 65535) // Valid CCSID and not binary
-    {
-      encoding_to_use = zut_int_to_string(file_ccsid);
-      has_encoding = true;
-    }
-    else if (strlen(zusf->encoding_opts.codepage) > 0)
+    if (strlen(zusf->encoding_opts.codepage) > 0)
     {
       encoding_to_use = string(zusf->encoding_opts.codepage);
       has_encoding = true;
+    }
+    else
+    {
+      // Use tagged encoding if valid CCSID and not UTF-8 or binary
+      int file_ccsid = zusf_get_file_ccsid(zusf, file);
+      if (file_ccsid > 0 && file_ccsid != 1208 && file_ccsid != 65535)
+      {
+        encoding_to_use = zut_int_to_string(file_ccsid);
+        has_encoding = true;
+      }
     }
   }
 
@@ -1291,12 +1314,19 @@ int zusf_write_to_uss_file(ZUSF *zusf, string file, string &data)
  * @param zusf pointer to a ZUSF object
  * @param file name of the USS file
  * @param pipe name of the input pipe
+ * @param content_len pointer where the length of the file contents will be stored
  *
  * @return RTNCD_SUCCESS on success, RTNCD_FAILURE on failure
  */
-int zusf_write_to_uss_file_streamed(ZUSF *zusf, string file, string pipe)
+int zusf_write_to_uss_file_streamed(ZUSF *zusf, const string &file, const string &pipe, size_t *content_len)
 {
   // TODO(zFernand0): Avoid overriding existing files
+  if (content_len == nullptr)
+  {
+    zusf->diag.e_msg_len = sprintf(zusf->diag.e_msg, "content_len must be a valid size_t pointer");
+    return RTNCD_FAILURE;
+  }
+
   struct stat file_stats;
 
   // Use file tag encoding if available, otherwise fall back to provided encoding
@@ -1305,17 +1335,20 @@ int zusf_write_to_uss_file_streamed(ZUSF *zusf, string file, string pipe)
 
   if (zusf->encoding_opts.data_type == eDataTypeText)
   {
-    // Try to get the file's CCSID first
-    int file_ccsid = zusf_get_file_ccsid(zusf, file);
-    if (file_ccsid > 0 && file_ccsid != 65535) // Valid CCSID and not binary
-    {
-      encoding_to_use = zut_int_to_string(file_ccsid);
-      has_encoding = true;
-    }
-    else if (strlen(zusf->encoding_opts.codepage) > 0)
+    if (strlen(zusf->encoding_opts.codepage) > 0)
     {
       encoding_to_use = string(zusf->encoding_opts.codepage);
       has_encoding = true;
+    }
+    else
+    {
+      // Use tagged encoding if valid CCSID and not UTF-8 or binary
+      int file_ccsid = zusf_get_file_ccsid(zusf, file);
+      if (file_ccsid > 0 && file_ccsid != 1208 && file_ccsid != 65535)
+      {
+        encoding_to_use = zut_int_to_string(file_ccsid);
+        has_encoding = true;
+      }
     }
   }
 
@@ -1357,6 +1390,7 @@ int zusf_write_to_uss_file_streamed(ZUSF *zusf, string file, string pipe)
     std::vector<char> temp_encoded = zbase64::decode(&buf[0], bytes_read);
     const char *chunk = &temp_encoded[0];
     int chunk_len = temp_encoded.size();
+    *content_len += chunk_len;
 
     if (has_encoding)
     {
@@ -1523,7 +1557,7 @@ const char *zusf_get_group_from_gid(gid_t gid)
   return meta ? meta->gr_name : nullptr;
 }
 
-short zusf_get_id_from_user_or_group(string user_or_group, bool is_user)
+short zusf_get_id_from_user_or_group(const string &user_or_group, bool is_user)
 {
   const auto is_numeric = user_or_group.find_first_not_of("0123456789") == std::string::npos;
   if (is_numeric)
@@ -1540,7 +1574,7 @@ short zusf_get_id_from_user_or_group(string user_or_group, bool is_user)
   return -1;
 }
 
-int zusf_chown_uss_file_or_dir(ZUSF *zusf, string file, string owner, bool recursive)
+int zusf_chown_uss_file_or_dir(ZUSF *zusf, string file, const string &owner, bool recursive)
 {
   struct stat file_stats;
   if (stat(file.c_str(), &file_stats) == -1)