+++ native/c/zusf.hpp
@@ -29,17 +29,17 @@ typedef struct _ListOptions
   bool long_format;
 } ListOptions;
 
-int zusf_create_uss_file_or_dir(ZUSF *zusf, std::string file, mode_t mode, bool createDir);
+int zusf_create_uss_file_or_dir(ZUSF *zusf, const std::string &file, mode_t mode, bool createDir);
 std::string zusf_format_file_entry(ZUSF *zusf, const struct stat &file_stats, const std::string &file_path, const std::string &display_name, ListOptions options, bool use_csv_format);
 int zusf_list_uss_file_path(ZUSF *zusf, std::string file, std::string &response, ListOptions options = ListOptions{}, bool use_csv_format = false);
-int zusf_read_from_uss_file(ZUSF *zusf, std::string file, std::string &response);
-int zusf_read_from_uss_file_streamed(ZUSF *zusf, std::string file, std::string pipe);
-int zusf_write_to_uss_file(ZUSF *zusf, std::string file, std::string &data);
-int zusf_write_to_uss_file_streamed(ZUSF *zusf, std::string file, std::string pipe);
+int zusf_read_from_uss_file(ZUSF *zusf, const std::string &file, std::string &response);
+int zusf_read_from_uss_file_streamed(ZUSF *zusf, const std::string &file, const std::string &pipe, size_t *content_len);
+int zusf_write_to_uss_file(ZUSF *zusf, const std::string &file, std::string &data);
+int zusf_write_to_uss_file_streamed(ZUSF *zusf, const std::string &file, const std::string &pipe, size_t *content_len);
 int zusf_chmod_uss_file_or_dir(ZUSF *zusf, std::string file, mode_t mode, bool recursive);
 int zusf_delete_uss_item(ZUSF *zusf, std::string file, bool recursive);
-int zusf_chown_uss_file_or_dir(ZUSF *zusf, std::string file, std::string owner, bool recursive);
-short zusf_get_id_from_user_or_group(std::string user_or_group, bool is_user);
+int zusf_chown_uss_file_or_dir(ZUSF *zusf, std::string file, const std::string &owner, bool recursive);
+short zusf_get_id_from_user_or_group(const std::string &user_or_group, bool is_user);
 int zusf_chtag_uss_file_or_dir(ZUSF *zusf, std::string file, std::string tag, bool recursive);
 int zusf_get_file_ccsid(ZUSF *zusf, std::string file);
 std::string zusf_get_ccsid_display_name(int ccsid);