+++ packages/cli/src/upload/file-to-uss/FileToUss.handler.ts
@@ -11,34 +11,15 @@
 
 import * as fs from "node:fs";
 import type { IHandlerParameters } from "@zowe/imperative";
-import { B64String, type ZSshClient, type uss } from "zowe-native-proto-sdk";
+import type { ZSshClient, uss } from "zowe-native-proto-sdk";
 import { SshBaseHandler } from "../../SshBaseHandler";
 
 export default class UploadFileToUssFileHandler extends SshBaseHandler {
     public async processWithClient(params: IHandlerParameters, client: ZSshClient): Promise<uss.WriteFileResponse> {
-        let encoding = params.arguments.encoding;
-        const binary = params.arguments.binary;
-        if (encoding == null && binary == null) {
-            try {
-                const fileResp = await client.uss.listFiles({
-                    fspath: params.arguments.ussFile,
-                    all: true,
-                    long: true,
-                });
-                if (fileResp.success && fileResp.items.length > 0) {
-                    const file = fileResp.items[0];
-                    encoding = file.filetag;
-                }
-            } catch (error) {
-                // Ignore as the file may not exist yet
-            }
-        }
-
-        const data = B64String.encode(fs.readFileSync(params.arguments.file));
         const response = await client.uss.writeFile({
-            data,
+            stream: fs.createReadStream(params.arguments.file),
             fspath: params.arguments.ussFile,
-            encoding: binary ? "binary" : encoding,
+            encoding: params.arguments.binary ? "binary" : params.arguments.encoding,
         });
         const uploadSource: string = `local file '${params.arguments.file}'`;
         const successMsg = params.response.console.log(